+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "rust-analyzer.linkedProjects": ["./Cargo.toml"],
+  "rust-analyzer.showUnlinkedFileNotification": false
+}
