+++ Cargo.toml
@@ -6,7 +6,5 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-package = "0.0.0"
-wgpu = "0.15.1"
+ggez = "0.9.3"
 winapi = {version = "0.3.9", features = ["everything"]}
-winit = "0.28.3"
