+++ src/core/memory_reader.rs
@@ -1,0 +1,3 @@
+use crate::helpers::MemoryHelper;
+
+pub struct MemoryReader {}
