+++ src/core/mod.rs
@@ -1,0 +1,2 @@
+mod memory_reader;
+pub use memory_reader::MemoryReader;
